Add dummy pubsub to stream random trades into db

+++ internal/pkg/db/migrations/0001_trades.sql
@@ -2,8 +2,8 @@
 CREATE TABLE trades (
   id serial PRIMARY KEY,
   created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  instrument_id int NOT NULL,
-  size int NOT NULL,
+  instrument_id bigint NOT NULL,
+  size bigint NOT NULL,
   price numeric NOT NULL,
   timestamp timestamp without time zone NOT NULL
 );
